REFACTOR: Rename API
